fix: fix update and delete file

+++ backend/auth-service/src/main/java/com/hotel/auth_service/service/impl/FileStorageServiceImpl.java
@@ -1,17 +1,18 @@
 package com.hotel.auth_service.service.impl;
 
-import com.hotel.auth_service.service.FileStorageService;
+import java.net.MalformedURLException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.nio.file.StandardCopyOption;
+
 import org.springframework.core.io.Resource;
 import org.springframework.core.io.UrlResource;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.net.MalformedURLException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.nio.file.StandardCopyOption;
+import com.hotel.auth_service.service.FileStorageService;
 
 @Service
 public class FileStorageServiceImpl implements FileStorageService {
@@ -66,7 +67,9 @@ public class FileStorageServiceImpl implements FileStorageService {
     public void deleteFile(String fileName) {
         try {
             Path filePath = this.fileStorageLocation.resolve(fileName).normalize();
-            Files.delete(filePath);
+            if (Files.exists(filePath)) {
+                Files.delete(filePath);
+            }
         } catch (Exception ex) {
             throw new RuntimeException("Could not delete file " + fileName, ex);
         }
